Tools update
VRF Decompiler v15.0 c5157 => v16.0 c5262
SDK Source2Gen c265 => c265 pr100
radare2 v6.0.4 => v6.0.7
ProtobufDumper c4064 => c4081
DepotDownloader v3.4.0 c593 => v3.4.0 c605

+++ sdk/include/source2sdk/modellib/InputLayoutVariation_t.hpp
@@ -14,9 +14,9 @@ namespace source2sdk
     namespace modellib
     {
         // Enumerator count: 4
-        // Alignment: 4
+        // Alignment: 1
         // Size: 0x_
-        enum class InputLayoutVariation_t : std::uint32_t
+        enum class InputLayoutVariation_t : std::uint8_t
         {
             INPUT_LAYOUT_VARIATION_DEFAULT = 0x0,
             INPUT_LAYOUT_VARIATION_STREAM1_INSTANCEID = 0x1,
